Revert "1.19"
