+++ gradle.properties
@@ -6,7 +6,7 @@ minecraft_version=1.19.2
 yarn_mappings=1.19.2+build.18
 loader_version=0.14.9
 # Mod Properties
-mod_version=0.1.6
+mod_version=0.1.4
 maven_group=com
 archives_base_name=miskatonicmysteries
 # Dependencies
