+++ src/main/java/com/miskatonicmysteries/mixin/block/BlockMixin.java
@@ -4,7 +4,9 @@ import com.miskatonicmysteries.api.MiskatonicMysteriesAPI;
 import com.miskatonicmysteries.common.MMMidnightLibConfig;
 import com.miskatonicmysteries.common.handler.InsanityHandler;
 import com.miskatonicmysteries.common.handler.networking.packet.c2s.InvokeManiaPacket;
+import com.miskatonicmysteries.common.registry.MMObjects;
 import com.miskatonicmysteries.common.util.Constants;
+import com.miskatonicmysteries.common.util.Util;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
