+++ src/main/java/com/miskatonicmysteries/mixin/client/LightmapTextureManagerMixin.java
@@ -7,6 +7,7 @@ import net.fabricmc.api.Environment;
 
 import net.minecraft.client.render.LightmapTextureManager;
 import net.minecraft.util.math.MathHelper;
+import net.minecraft.world.World;
 
 import net.minecraft.world.dimension.DimensionType;
 import org.spongepowered.asm.mixin.Mixin;
