--- src/main/resources/data/minecraft/tags/point_of_interest_type/acquirable_job_site.json
@@ -1,6 +1,0 @@
-{
-  "replace": false,
-  "values": [
-    "miskatonicmysteries:psycho_poi"
-  ]
-}
